Issue #394 Move and clean up GettingStarted * Test not called out have been moved to relevant test classes
