+++ impl/src/test/java/org/ehcache/internal/store/heap/OnHeapStoreByValueTest.java
@@ -15,10 +15,15 @@
  */
 package org.ehcache.internal.store.heap;
 
+import org.ehcache.Cache;
+import org.ehcache.CacheManager;
+import org.ehcache.CacheManagerBuilder;
+import org.ehcache.config.CacheConfigurationBuilder;
 import org.ehcache.config.Eviction;
 import org.ehcache.config.EvictionVeto;
 import org.ehcache.config.EvictionPrioritizer;
 import org.ehcache.config.ResourcePools;
+import org.ehcache.config.ResourcePoolsBuilder;
 import org.ehcache.config.units.EntryUnit;
 import org.ehcache.exceptions.SerializerException;
 import org.ehcache.expiry.Expirations;
@@ -26,6 +31,7 @@ import org.ehcache.expiry.Expiry;
 import org.ehcache.internal.SystemTimeSource;
 import org.ehcache.internal.TimeSource;
 import org.ehcache.internal.serialization.JavaSerializer;
+import org.ehcache.internal.store.heap.service.OnHeapStoreServiceConfiguration;
 import org.ehcache.spi.cache.Store;
 import org.ehcache.spi.cache.Store.ValueHolder;
 import org.junit.Test;
@@ -36,6 +42,8 @@ import java.util.Collections;
 import java.util.List;
 
 import static org.ehcache.config.ResourcePoolsBuilder.newResourcePoolsBuilder;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.fail;
 
 @SuppressWarnings("serial")
@@ -101,6 +109,29 @@ public class OnHeapStoreByValueTest extends BaseOnHeapStoreTest {
     }
   }
 
+  @Test
+  public void testStoreByValue() {
+    CacheManager cacheManager = CacheManagerBuilder.newCacheManagerBuilder().build(false);
+    cacheManager.init();
+
+    final Cache<Long, String> cache1 = cacheManager.createCache("cache1",
+        CacheConfigurationBuilder.newCacheConfigurationBuilder().withResourcePools(ResourcePoolsBuilder.newResourcePoolsBuilder().heap(1, EntryUnit.ENTRIES))
+            .buildConfig(Long.class, String.class));
+    performAssertions(cache1, true);
+
+    final Cache<Long, String> cache2 = cacheManager.createCache("cache2",
+        CacheConfigurationBuilder.newCacheConfigurationBuilder().add(new OnHeapStoreServiceConfiguration().storeByValue(true))
+            .buildConfig(Long.class, String.class));
+    performAssertions(cache2, false);
+
+    final Cache<Long, String> cache3 = cacheManager.createCache("cache3",
+        CacheConfigurationBuilder.newCacheConfigurationBuilder().add(new OnHeapStoreServiceConfiguration().storeByValue(false))
+            .buildConfig(Long.class, String.class));
+    performAssertions(cache3, true);
+
+    cacheManager.close();
+  }
+
   @Override
   protected <K, V> OnHeapStore<K, V> newStore() {
     return newStore(SystemTimeSource.INSTANCE, Expirations.noExpiration(), Eviction.none());
@@ -160,4 +191,13 @@ public class OnHeapStoreByValueTest extends BaseOnHeapStoreTest {
     }, timeSource, true, new JavaSerializer<K>(getClass().getClassLoader()), new JavaSerializer<V>(getClass().getClassLoader()));
   }
 
+  private void performAssertions(Cache<Long, String> cache, boolean same) {
+    cache.put(1L, "one");
+    String s1 = cache.get(1L);
+    String s2 = cache.get(1L);
+    String s3 = cache.get(1L);
+
+    assertThat(s1 == s2, is(same));
+    assertThat(s2 == s3, is(same));
+  }
 }
